Try to be as flexible/supportive and future proof
Send also timeout rcon command in addition to ZillyWoods messages.
In case the old ddnet system will be reintroduced it will me a chat command and thus also a rcon command.
So the server code can be cleaner just supporting the rcon. Which is better than the chat system anyways.